# openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Wavelet API",
		"version": "1.0.0",
		"description": "Music streaming / library API behind wavelet (Next.js on Vercel).\n\n## Response envelope\nEvery `/api/v1/*` JSON route returns `{ \"success\": true, \"data\": … }` or `{ \"success\": false, \"error\": { \"code\", \"message\" } }`.\nBranch on `error.code`, not on the message. `/api/auth/*` (better-auth) does NOT use this envelope.\n\n## Authentication\nbetter-auth session (Google only), accepted two ways:\n- **Bearer (native / Flutter)**: sign in with `POST /api/auth/sign-in/social` + Google `idToken`, read the `set-auth-token` response header, then send `Authorization: Bearer <token>` on every request — including the audio player's requests. Sessions last 30 days.\n- **Cookie (web)**: `__Secure-better-auth.session_token` (plus `__Secure-better-auth.session_data`, a 5-min signed cache).\nMany routes also need a connected Deezer account (`NO_DEEZER_ARL` → call `/api/v1/auth/login-arl`).\n\n## Ids\n`trackId`, `deezerAlbumId`, `deezerArtistId` are Deezer ids (as strings). `albumId` in `/library/albums/{albumId}` and playlist ids are internal cuids."
	},
	"servers": [
		{
			"url": "https://wavelet.titosy.dev",
			"description": "Production"
		},
		{
			"url": "http://localhost:3000",
			"description": "Local dev (`npm run dev`)"
		}
	],
...
	"security": [
		{
			"bearerAuth": []
		},
		{
			"sessionCookie": []
		}
	],
	"paths": {
...
		"/api/v1/stream-warm/{trackId}": {
			"parameters": [
				{
					"name": "trackId",
					"in": "path",
					"required": true,
					"description": "Deezer track id",
					"schema": {
						"type": "string"
					}
				}
			],
			"get": {
				"tags": [
					"Streaming"
				],
				"operationId": "warmStream",
				"summary": "Prefetch Deezer metadata so the next play starts faster",
				"description": "Fire-and-forget, always 204 on success. Call when a track is about to be played (e.g. next in queue).",
				"security": [
					{
						"bearerAuth": []
					},
					{
						"sessionCookie": []
					}
				],
				"responses": {
					"204": {
						"description": "Accepted (no body)"
					},
					"401": {
						"$ref": "#/components/responses/NotAuthenticatedOrDeezerLoginFailed"
					},
					"403": {
						"$ref": "#/components/responses/NoDeezerArl"
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		}
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"description": "Session token from the `set-auth-token` header returned by `/api/auth/sign-in/social`."
			},
			"sessionCookie": {
				"type": "apiKey",
				"in": "cookie",
				"name": "__Secure-better-auth.session_token",
				"description": "better-auth session cookie (`better-auth.session_token` on http://localhost)."
			}
		},
		"responses": {
...
			"NotAuthenticatedOrDeezerLoginFailed": {
				"description": "`NOT_AUTHENTICATED`, `DEEZER_LOGIN_FAILED` or `LOGIN_FAILED`",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorResponse"
						}
					}
				}
			},
			"NoDeezerArl": {
				"description": "`NO_DEEZER_ARL` — user has no Deezer account connected (send them to login-arl)",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorResponse"
						}
					}
				}
			},
...
			"InternalError": {
				"description": "`INTERNAL_ERROR`, `AUTH_ERROR`, `DEEZER_ERROR` or `APP_NOT_INITIALIZED`",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorResponse"
						}
					}
				}
			}
		},
		"schemas": {
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						]
					},
					"error": {
						"type": "object",
						"properties": {
							"code": {
								"type": "string",
								"description": "Machine-readable error code (see `x-error-codes`).",
								"example": "NOT_AUTHENTICATED"
							},
							"message": {
								"type": "string",
								"example": "Please sign in to continue."
							}
						},
						"required": [
							"code",
							"message"
						]
					}
				},
				"required": [
					"success",
					"error"
				]
			},
...
		}
	}
}